Create README.md
This website has java. I have a particular set of skills. Coding as I've come to learn is not one of them. While it has been enjoyable to experience and work through, all concepts are hard for me to execute. Very simply I coded some animation of a line traveling across the canvas at framerate of (1345) and manipulated the oscillator. I was trying to make the gif on the left of the site the entire background, but it didn't seem to work as I looked at my old style sheets where it had worked.